Move level and prefixer into separate files
